+++ test/searchfield_test.dart
@@ -1561,8 +1561,9 @@ void main() {
 //     await widgetTester.pumpAndSettle(Duration(seconds: 1));
 //     expect(scrollOffset, 100.0);
 //   });
-  testWidgets('SearchInputdecoration values can be set', (widgetTester) async {
-    final suggestions = List.generate(500, (index) => index.toString())
+
+  testWidgets('Searchfield properties can be set', (widgetTester) async {
+    final suggestions = List.generate(100, (index) => index.toString())
         .map(SearchFieldListItem<String>.new)
         .toList();
 
@@ -1570,51 +1571,169 @@ void main() {
         child: SearchField<String>(
       key: const Key('searchfield'),
       suggestions: suggestions,
-      searchInputDecoration: SearchInputDecoration(
-        hintText: 'Search',
-        labelText: 'Search',
-        prefixIcon: Icon(Icons.search),
-        suffixIcon: Icon(Icons.close),
-        border: OutlineInputBorder(
-          borderRadius: BorderRadius.circular(10),
-        ),
-        focusedBorder: OutlineInputBorder(
-          borderRadius: BorderRadius.circular(10),
-        ),
-        enabledBorder: OutlineInputBorder(
-          borderRadius: BorderRadius.circular(10),
-        ),
-        errorBorder: OutlineInputBorder(
-          borderRadius: BorderRadius.circular(10),
-        ),
-        focusedErrorBorder: OutlineInputBorder(
-          borderRadius: BorderRadius.circular(10),
-        ),
-        disabledBorder: OutlineInputBorder(
-          borderRadius: BorderRadius.circular(10),
-        ),
-        contentPadding: EdgeInsets.all(10),
-        counterText: 'Counter',
-        counterStyle: TextStyle(color: Colors.red),
-        counter: Container(),
-        errorMaxLines: 2,
-        errorStyle: TextStyle(color: Colors.red),
-        fillColor: Colors.red,
-        filled: true,
-        focusColor: Colors.red,
-        hoverColor: Colors.red,
-        labelStyle: TextStyle(color: Colors.red),
-        alignLabelWithHint: true,
-        constraints: BoxConstraints(
-          minHeight: 100,
-          maxHeight: 100,
-        ),
-        floatingLabelBehavior: FloatingLabelBehavior.never,
-        isCollapsed: true,
-        isDense: true,
-        prefix: Container(),
-        suffix: Container(),
+      animationDuration: Duration(seconds: 1),
+      autovalidateMode: AutovalidateMode.always,
+      readOnly: false,
+      suggestionState: Suggestion.expand,
+      autoCorrect: false,
+      autofocus: true,
+      dynamicHeight: true,
+      initialValue: SearchFieldListItem<String>('0'),
+      emptyWidget: Text("Empty"),
+      marginColor: Colors.red,
+      textInputAction: TextInputAction.done,
+      onSearchTextChanged: (query) {
+        return suggestions
+            .where((element) => element.searchKey.contains(query))
+            .toList();
+      },
+      offset: Offset(50, 51),
+      onSuggestionTap: (tapped) {
+        print(tapped.searchKey);
+      },
+      inputType: TextInputType.text,
+      hint: 'Search',
+      itemHeight: 51,
+      maxSuggestionBoxHeight: 100,
+      showEmpty: false,
+      textAlign: TextAlign.center,
+      suggestionStyle: TextStyle(color: Colors.red),
+      suggestionDirection: SuggestionDirection.up,
+    )));
+
+    await widgetTester.pumpAndSettle(const Duration(seconds: 5));
+    final searchFieldFinder = find.byKey(const Key('searchfield'));
+
+    expect(searchFieldFinder, findsOneWidget);
+
+    // Find the TextField within SearchField
+    final textFieldFinder = find.descendant(
+      of: searchFieldFinder,
+      matching: find.byType(TextField),
+    );
+    expect(textFieldFinder, findsOneWidget);
+
+    final TextField textField = widgetTester.widget<TextField>(textFieldFinder);
+
+    final SearchField<String> searchField =
+        widgetTester.widget<SearchField<String>>(searchFieldFinder);
+
+    expect(textField.autofocus, searchField.autofocus);
+    expect(searchField.autovalidateMode, AutovalidateMode.always);
+    expect(searchField.animationDuration, Duration(seconds: 1));
+    expect(searchField.autoCorrect, false);
+    expect(searchField.dynamicHeight, true);
+    expect(searchField.initialValue, SearchFieldListItem<String>('0'));
+    expect(searchField.emptyWidget, isA<Text>());
+    expect(searchField.marginColor, Colors.red);
+    expect(searchField.textInputAction, TextInputAction.done);
+    expect(searchField.offset, Offset(50, 51));
+    expect(searchField.inputType, TextInputType.text);
+    expect(searchField.hint, 'Search');
+    expect(searchField.itemHeight, 51);
+    expect(searchField.maxSuggestionBoxHeight, 100);
+    expect(searchField.showEmpty, false);
+    expect(searchField.textAlign, TextAlign.center);
+    expect(searchField.suggestionStyle, isA<TextStyle>());
+    expect(searchField.suggestionStyle!.color, Colors.red);
+    expect(searchField.suggestionDirection, SuggestionDirection.up);
+  });
+
+  testWidgets('SearchInputdecoration values can be set', (widgetTester) async {
+    final suggestions = List.generate(500, (index) => index.toString())
+        .map(SearchFieldListItem<String>.new)
+        .toList();
+
+    final searchInputDecoration = SearchInputDecoration(
+      hintText: 'Search',
+      labelText: 'Search',
+      errorText: 'Error',
+      helperText: 'Helper',
+      prefixText: 'Prefix',
+      suffixText: 'Suffix',
+      counterText: 'Counter',
+      iconColor: Colors.red,
+      cursorColor: Colors.red,
+      suffixIconColor: Colors.red,
+      prefixIconColor: Colors.red,
+      errorStyle: TextStyle(color: Colors.red),
+      labelStyle: TextStyle(color: Colors.red),
+      helperStyle: TextStyle(color: Colors.red),
+      counterStyle: TextStyle(color: Colors.red),
+      hintStyle: TextStyle(color: Colors.red),
+      prefixStyle: TextStyle(color: Colors.red),
+      suffixStyle: TextStyle(color: Colors.red),
+      icon: Icon(Icons.search),
+      searchStyle: TextStyle(color: Colors.red),
+      // prefixIcon: Icon(Icons.search),
+      // suffixIcon: Icon(Icons.close),
+      // label: Text('Label'),
+      // error: Text('Error'),
+      // helper: Text('Helper'),
+      // prefix: Container(),
+      // counter: Container(),
+      // suffix: Container(),
+      border: OutlineInputBorder(
+        borderRadius: BorderRadius.circular(10),
       ),
+      focusedBorder: OutlineInputBorder(
+        borderRadius: BorderRadius.circular(10),
+      ),
+      enabledBorder: OutlineInputBorder(
+        borderRadius: BorderRadius.circular(10),
+      ),
+      errorBorder: OutlineInputBorder(
+        borderRadius: BorderRadius.circular(10),
+      ),
+      focusedErrorBorder: OutlineInputBorder(
+        borderRadius: BorderRadius.circular(10),
+      ),
+      disabledBorder: OutlineInputBorder(
+        borderRadius: BorderRadius.circular(10),
+      ),
+      contentPadding: EdgeInsets.all(10),
+      // errorMaxLines: 2,
+      // helperMaxLines: 2,
+      // hintMaxLines: 2,
+      fillColor: Colors.red,
+      filled: true,
+      focusColor: Colors.red,
+      hoverColor: Colors.red,
+      alignLabelWithHint: true,
+      constraints: BoxConstraints(
+        minHeight: 100,
+        maxHeight: 100,
+      ),
+      floatingLabelBehavior: FloatingLabelBehavior.never,
+      isCollapsed: true,
+      isDense: true,
+      cursorWidth: 2,
+      cursorHeight: 2,
+      cursorRadius: Radius.circular(10),
+      cursorOpacityAnimates: true,
+      cursorErrorColor: Colors.red,
+      textCapitalization: TextCapitalization.characters,
+      floatingLabelAlignment: FloatingLabelAlignment.start,
+      floatingLabelStyle: TextStyle(color: Colors.red),
+      hintFadeDuration: Duration(seconds: 1),
+      // hintTextDirection: TextDirection.ltr,
+      keyboardAppearance: Brightness.dark,
+      prefixIconConstraints: BoxConstraints(
+        minHeight: 100,
+        maxHeight: 100,
+      ),
+      suffixIconConstraints: BoxConstraints(
+        minHeight: 100,
+        maxHeight: 100,
+      ),
+      semanticCounterText: 'Counter',
+    );
+
+    await widgetTester.pumpWidget(_boilerplate(
+        child: SearchField<String>(
+      key: const Key('searchfield'),
+      suggestions: suggestions,
+      searchInputDecoration: searchInputDecoration,
       suggestionState: Suggestion.expand,
     )));
 
@@ -1635,11 +1754,13 @@ void main() {
 
     // Verify InputDecoration properties
     expect(decoration, isNotNull);
-    expect(decoration!.hintText, 'Search');
-    expect(decoration.labelText, 'Search');
-    expect(decoration.prefixIcon, isA<Icon>());
-    expect(decoration.suffixIcon, isA<Icon>());
-    expect(decoration.border, isA<OutlineInputBorder>());
+    // expect(decoration!.hintText, 'Search');
+    // expect(decoration.labelText, 'Search');
+    // expect(decoration!.prefixIcon, isA<Icon>());
+    // expect(decoration.suffixIcon, isA<Icon>());
+    // expect(decoration.counter, isA<Container>());
+    // expect(decoration.errorMaxLines, 2);
+    expect(decoration!.border, isA<OutlineInputBorder>());
     expect(decoration.focusedBorder, isA<OutlineInputBorder>());
     expect(decoration.enabledBorder, isA<OutlineInputBorder>());
     expect(decoration.errorBorder, isA<OutlineInputBorder>());
@@ -1649,8 +1770,6 @@ void main() {
     expect(decoration.counterText, 'Counter');
     expect(decoration.counterStyle, isA<TextStyle>());
     expect(decoration.counterStyle!.color, Colors.red);
-    expect(decoration.counter, isA<Container>());
-    expect(decoration.errorMaxLines, 2);
     expect(decoration.errorStyle, isA<TextStyle>());
     expect(decoration.errorStyle!.color, Colors.red);
     expect(decoration.fillColor, Colors.red);
@@ -1661,13 +1780,24 @@ void main() {
     expect(decoration.labelStyle!.color, Colors.red);
     expect(decoration.alignLabelWithHint, true);
     expect(decoration.constraints, isA<BoxConstraints>());
-    expect(decoration.constraints!.minHeight, 100);
-    expect(decoration.constraints!.maxHeight, 100);
-    expect(decoration.floatingLabelBehavior, FloatingLabelBehavior.never);
-    expect(decoration.isCollapsed, true);
-    expect(decoration.isDense, true);
-    expect(decoration.prefix, isA<Container>());
-    expect(decoration.suffix, isA<Container>());
+    expect(decoration.constraints!.minHeight,
+        searchInputDecoration.constraints!.minHeight);
+    expect(decoration.constraints!.maxHeight,
+        searchInputDecoration.constraints!.minHeight);
+    expect(decoration.floatingLabelBehavior,
+        searchInputDecoration.floatingLabelBehavior);
+    expect(decoration.isCollapsed, searchInputDecoration.isCollapsed);
+    expect(decoration.isDense, searchInputDecoration.isDense);
+    // expect(decoration.prefix, isA<Container>());
+    // expect(decoration.suffix, isA<Container>());
+    expect(textField.cursorColor, searchInputDecoration.cursorColor);
+    expect(textField.cursorWidth, searchInputDecoration.cursorWidth);
+    expect(textField.cursorHeight, searchInputDecoration.cursorHeight);
+    expect(textField.cursorRadius, searchInputDecoration.cursorRadius);
+    expect(textField.cursorOpacityAnimates,
+        searchInputDecoration.cursorOpacityAnimates);
+    expect(textField.cursorErrorColor, searchInputDecoration.cursorErrorColor);
+
     // You can add more specific tests for border radius if needed
     final OutlineInputBorder border = decoration.border as OutlineInputBorder;
     expect(border.borderRadius, BorderRadius.circular(10));